feat: add details block
