+++ src/Content.tsx
@@ -1,13 +1,15 @@
 import ProductCard from './ProductCard';
 import Paginator from './Paginator';
 import SelectPages from './SelectPages';
-import { useEffect, useRef } from 'react';
+import { useEffect, useRef, useState } from 'react';
+import { Outlet, useNavigate } from 'react-router-dom';
 
 interface ContentProps {
   products: Array<{
     title: string;
     thumbnail: string;
     description: string;
+    id: number;
   }> | null;
   isButtonDisabled: boolean;
   setCardsPerPage: (string: string) => void;
@@ -24,7 +26,17 @@ function Content({
   totalProducts,
   handleClick,
 }: ContentProps) {
+  const navigate = useNavigate();
   const isMounted = useRef(false);
+  const [detailsAreClosed, setDetailsAreClosed] = useState(true);
+  const handleContentClick = () => {
+    if (!detailsAreClosed) {
+      navigate('../');
+      setDetailsAreClosed(true);
+    }
+
+    console.log(detailsAreClosed);
+  };
   useEffect(() => {
     if (isMounted.current) {
       handleClick('1');
@@ -43,11 +55,16 @@ function Content({
   if (products && products.length > 0) {
     return (
       <main className="main">
-        <SelectPages setCardsPerPage={setCardsPerPage} />
-        <div className="products-data">
-          {products.map((product, index) => (
-            <ProductCard key={index} product={product} />
-          ))}
+        <div className="content-wrapper" onClick={handleContentClick}>
+          <SelectPages setCardsPerPage={setCardsPerPage} />
+          <div className="products">
+            <div className="products-data" onClick={handleContentClick}>
+              {products.map((product, index) => (
+                <ProductCard key={index} product={product} />
+              ))}
+            </div>
+            <Outlet />
+          </div>
         </div>
         <Paginator
           cardsPerPage={cardsPerPage}
