+++ src/_index.scss
@@ -191,3 +191,15 @@ h1 {
     transform: rotate(360deg);
   }
 }
+
+.products {
+  display: flex;
+  &-details {
+    border: 1px solid black;
+    min-width: 400px;
+    img {
+      width: 100%;
+      height: auto;
+    }
+  }
+}
